add cocktail menu server template
