+++ lib/pdf-server-components/pdf-template-factory.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import { templateRegistry, TemplateMetadata } from './template-registry'
-import { MenuEditorProvider } from '@/contexts/menu-editor-context'
 
 // Global flag to indicate PDF generation mode
 let isPDFGenerationMode = false
@@ -108,18 +107,13 @@ export class PDFTemplateFactory {
       // Load template component dynamically  
       const TemplateComponent = await templateRegistry.loadTemplateComponent(normalizedId)
       
-      // Create the template element with a wrapper that provides context data
-      const templateElement = React.createElement(
-        PDFTemplateContextProvider,
-        {
-          restaurant: props.restaurant,
-          categories: props.categories,
-          language: props.language,
-          customizations: props.customizations,
-          children: React.createElement(TemplateComponent)
-        }
-      )
-      
+      // Directly render the template component with supplied props
+      const templateElement = React.createElement(TemplateComponent, {
+        ...props,
+        isPdfGeneration: true,
+        isPreview: false,
+      })
+
       return templateElement
     } catch (error) {
       console.error(`❌ Error creating template ${templateId}:`, error)
@@ -264,47 +258,4 @@ export const DynamicPDFTemplate: React.FC<{
   return React.createElement(TemplateComponent, props)
 }
 
-/**
- * PDF Context provider that wraps templates with MenuEditorProvider
- * This ensures templates get all the context data they need
- */
-const PDFTemplateContextProvider: React.FC<{
-  restaurant: Restaurant
-  categories: MenuCategory[]
-  language?: string
-  customizations?: any
-  children: React.ReactNode
-}> = ({ restaurant, categories, language, customizations, children }) => {
-  // Create a proper MenuEditorProvider with the PDF data
-  const contextRestaurant = {
-    ...restaurant,
-    // Ensure color_palette has the right structure for context
-    color_palette: restaurant.color_palette ? {
-      id: 'pdf-palette',
-      name: 'PDF Palette',
-      ...restaurant.color_palette
-    } : null,
-    // Ensure address/phone/website are properly typed
-    address: restaurant.address || undefined,
-    phone: restaurant.phone || undefined,
-    website: restaurant.website || undefined,
-  }
-
-  return React.createElement(
-    MenuEditorProvider,
-    {
-      restaurant: contextRestaurant as any, // Type cast to avoid complex type issues
-      initialCategories: categories,
-      onRefresh: () => {}, // No-op for PDF generation
-      children: React.createElement('div', { 
-        'data-pdf-mode': 'true',
-        style: { 
-          width: '100%', 
-          height: '100%' 
-        }
-      }, children)
-    }
-  )
-}
-
-export default PDFTemplateFactory 
+export default PDFTemplateFactory
